Add a more comprehensive JSON file to track versions (#101)
A fully automated image publishing pipeline needs to access a file that tells it exactly which products and components to update, and to what version. In contrast, our current pipeline requires that someone supply the version as a parameter when they run the pipeline. Not only is this error prone, but it is inflexible--we have a single version parameter that applies to all images.

We have some existing files that are obvious candidates for this purpose, but they each have problems:

[versionInfo.json](https://github.com/Azure/iotedge/blob/main/versionInfo.json) in the Azure/iotedge repository has a single version property, so it can’t describe several independently versioned components. It is currently used by Edge Agent and Edge Hub code to log the commit ID and version number at runtime. It is also used in some of our newer release automation to verify Docker tags, automatically update `latest-iotedge-lts.json` in our Azure/azure-iotedge repository and generate our GitHub release pages.

[latest-aziot-edge.json](https://github.com/Azure/azure-iotedge/blob/main/latest-aziot-edge.json) and [latest-aziot-identity-service.json](https://github.com/Azure/azure-iotedge/blob/main/latest-aziot-identity-service.json) in the Azure/azure-iotedge repository contain version numbers specific to edgelet and the identity service, so they can’t describe Docker images.   They are used by `iotedge check` to detect the latest version of IoT Edge.

[latest-iotedge-lts.json](https://github.com/Azure/azure-iotedge/blob/main/latest-iotedge-lts.json) in the Azure/azure-iotedge repository contains the edgelet version number. It is used by `iotedge check` to detect the latest version of IoT Edge 1.1 LTS. It also lists version numbers for Edge Agent and Edge Hub images, but they are unused. Its name and contents tie it directly to 1.1 LTS.

This update introduces a new file, _product-versions.json_. The Azure/azure-iotedge repository is the right place for this file because it is the home for files related to the IoT Edge _product_ (the thing we package, ship and support), as opposed to the IoT Edge _project_, which is where our open-source code lives.

This new file is versioned and has detailed information about the products and components we ship, their versions, and their supported release channels. It can be consumed immediately by automated image publishing pipelines. Existing tools (e.g., `iotedge check` and existing automation) can migrate as needed over time.